Migrated test functions to dbtest.go
